+++ nextcloudappstore/core/user/urls.py
@@ -1,10 +1,14 @@
 from django.conf.urls import url
 
 from nextcloudappstore.core.user.views import PasswordView, AccountView, \
-    APITokenView, DeleteAccountView, ChangeLanguageView
+    APITokenView, DeleteAccountView, ChangeLanguageView, TransferAppsView
 
 urlpatterns = [
     url(r'^$', AccountView.as_view(), name='account'),
+    url(r'^transfer-apps/?$', TransferAppsView.as_view(),
+        name='account-transfer-apps'),
+    url(r'^transfer-apps/(?P<pk>[a-z0-9_]+)/?$', TransferAppsView.as_view(),
+        name='account-transfer-app'),
     url(r'^password/?$', PasswordView.as_view(), name='account-password'),
     url(r'^token/?$', APITokenView.as_view(), name='account-api-token'),
     url(r'^delete/?$', DeleteAccountView.as_view(), name='account-deletion'),